Some adjustments to amqp producer
Since we only really support sending to
one destination at a time, either an exchange
or a queue, make those clearer both in the
interface and in the implementations.

Might think of actually receiving exchange and
routing key in the Publish function instead, so
it's more flexible, but will leave that for a
future refactor.

Also allow configuring published messages as
persistent, if they need to. I believe that
should be the case for events/webhooks, but
probably not necessarily for stream health logs.