Create SSE endpoint to subscribe to live events (#3)
* Make a couple changes after latest decisions

Mostly from livepeer-infra stuff

Changed some env vars to be easier to use
(for rabbitmq uris)

Changed sharding logic. Supports an explicit
shardPrefixes flag for now.

In that changed the way some of those stuff are
dealt with

* Create separate data package with event types

* Create API in healthcore to get past events

* Create API to get past + subscribe to new events

* Sort and crop past events slice

* Add an ID field to events

* Allow replaying SSE stream from a certain event

* Use httprouter lib for route matching

Seems to be the fastest one on top of net/http
To get some higher throughput, we can eventually
migrate to fasthttp.

* Create SSE API to subscribe to live events

* Bind to broadcater events specifically

* Handle shards comma-separate list on main

* Ignore error with mustFindLast query

It's just a nit anyway to allow users
to opt-out from a default error handlingg.

* Create RabbitMQ service when running locally

* Fix the URI parsing once and for all

The rabbitmq lib is very limited in terms
of what it does parsing the URIs, so we
both need some logic ourselves to handle
some things like defaults, as well as to
generate the amqp uri from the rabbitmq and
vice-versa.

* Grab updates from streams client

* analyzer: Create healtcheck API

* Makefile: Fix docker_run command

* pkg/event: Increase publish channel size

* api: Allow JSON long-polling on /events API

* Move the JSON-SSE logic to the (j)sse package

That will make it really easy to replicate the
protocol for other APIs, like one for only receiving
errors, metrics, etc.

* cmd/analyzer: Bind all flags with FlagVar funcs

We need some of them like that, but there was no
real reason to force the ones we don't to be regular
Flag() pointers. Consistency feels good.

* cmd/analyzer: Make cli flags kebab-case

This will create nicer envs with _ separators.

* health: Create logic to clean-up old streams data

* cmd/analyzer: Configure memory records TTL