health: Use pointers of status to avoid racy copies
@darkdarkdragon pointed out on https://github.com/livepeer/livepeer-data/issues/6
that a copy is not atomic so could have issues when copying a value
without holding a lock. I don't want to increaase the usage of the lock
either, so decided to change the status to a pointer instead so we will
have a consistent view of the status, not one in between a copy.

Also, this is probably better for perf anyway since we were copying that
big object multiple times on every event processing. Better like this I
guess.