pkg/event: Update bindings one existing streams
Now we'll make our first deploy already with
a change to the queue bindings. It already feels
overkill to have to change the stream name to do
that, so I'm  changing the logic on stream_consumer
to always set the bindings on the existing stream,
even if it already existed before. This works more
similarly to how other AMQP clients generally do
this.