api: Simplify authorization server interface
We just forward almost all of the full request, including
the original request method, and just set 2 additional headers
with the original URI that was called and the Livepeer Stream ID.

This mimics the auth-server logic from nginx a bit, and after
thinking about it I think it doesn't even make sense to make
a client out of it. Unless we want to later expose it for other
projects to re-use this logic, then we can create it I guess.