health: Start stream health state as nil
Instead of defaulting to false (unhealthy) state
while data for calculating stream health is not
available, just keep it as nil which should mean
"unknown".

This fixes the UI glitch on the beginning of a stream
which shows it as unhealthy until we receive the first
transcode event about it.